+++ Sources/BungoManifest/Protocols/BungoManifestSQLClient/SQLiteClient.swift
@@ -8,13 +8,13 @@ final class SQLiteClient: BungoManifestSQLClient {
         db = try Connection(path, readonly: true)
     }
 
-    func select(column: String, from table: String, where idColumn: String, is searchID: Int64) throws -> Data? {
+    func select(column: String, from table: String, where idColumn: String, is searchID: Int32) throws -> Data? {
         let table = Table(table)
         let id = Expression<Int64>(idColumn)
         let resultColumn = Expression<String>(column)
 
         let query = table.select(id, resultColumn)
-            .filter(id == searchID)
+            .filter(id == Int64(searchID))
 
         let items: AnySequence<Row>
